#235 - Add startup task - Create a new `GuildChecker` task that runs on startup and subsequently every 12 hours - Remove most things from the `on_ready` client event - Make GuildChecker wait for `on_ready` to be done - Make all the other tasks wait for GuildChecker to be complete too - Fix some of the logic in startup functions so it works as expected
