+++ discordbot/tasks/annualawards.py
@@ -8,10 +8,11 @@ from discordbot.stats.awardsbuilder import AwardsBuilder
 
 
 class AnnualBSEddiesAwards(commands.Cog):
-    def __init__(self, bot: discord.Client, guilds, logger):
+    def __init__(self, bot: discord.Client, guilds, logger, startup_tasks):
         self.bot = bot
         self.logger = logger
         self.guilds = guilds
+        self.startup_tasks = startup_tasks
 
         self.annual_bseddies_awards.start()
 
@@ -22,8 +23,20 @@ class AnnualBSEddiesAwards(commands.Cog):
         """
         self.annual_bseddies_awards.cancel()
 
+    def _check_start_up_tasks(self) -> bool:
+        """
+        Checks start up tasks
+        """
+        for task in self.startup_tasks:
+            if not task.finished:
+                return False
+        return True
+
     @tasks.loop(minutes=60)
     async def annual_bseddies_awards(self):
+        if not self._check_start_up_tasks():
+            self.logger.info("Startup tasks not complete - skipping loop")
+            return
         now = datetime.datetime.now()
 
         if not now.day == 2 or not now.hour == 14 or not now.month == 1:
