+++ discordbot/tasks/eddiegains.py
@@ -21,10 +21,11 @@ class EddieGainMessager(commands.Cog):
     generated by a cronjob and messages the individuals. It only messages them if they have the
     "The Boys" role.
     """
-    def __init__(self, bot: discord.Client, guilds, logger):
+    def __init__(self, bot: discord.Client, guilds, logger, startup_tasks):
         self.bot = bot
         self.guilds = guilds
         self.logger = logger
+        self.startup_tasks = startup_tasks
         self.user_points = UserPoints()
 
         self.eddie_manager = BSEddiesManager(self.bot, self.logger)
@@ -38,6 +39,15 @@ class EddieGainMessager(commands.Cog):
         """
         self.eddie_distributer.cancel()
 
+    def _check_start_up_tasks(self) -> bool:
+        """
+        Checks start up tasks
+        """
+        for task in self.startup_tasks:
+            if not task.finished:
+                return False
+        return True
+
     @tasks.loop(minutes=1)
     async def eddie_distributer(self):
         """
@@ -46,6 +56,9 @@ class EddieGainMessager(commands.Cog):
         daily BSEddies gain.
         :return:
         """
+        if not self._check_start_up_tasks():
+            self.logger.info("Startup tasks not complete - skipping loop")
+            return
         now = datetime.datetime.now()
 
         if now.hour != 7 or now.minute != 30:
