+++ discordbot/tasks/wordlereminder.py
@@ -8,12 +8,22 @@ from mongo.bsepoints import UserInteractions
 
 
 class WordleReminder(commands.Cog):
-    def __init__(self, bot: discord.Client, guilds, logger):
+    def __init__(self, bot: discord.Client, guilds, logger, startup_tasks):
         self.bot = bot
         self.logger = logger
         self.guilds = guilds
+        self.startup_tasks = startup_tasks
         self.user_interactions = UserInteractions()
         self.wordle_reminder.start()
+
+    def _check_start_up_tasks(self) -> bool:
+        """
+        Checks start up tasks
+        """
+        for task in self.startup_tasks:
+            if not task.finished:
+                return False
+        return True
 
     def cog_unload(self):
         """
@@ -28,6 +38,9 @@ class WordleReminder(commands.Cog):
         Loop that makes sure the King is assigned correctly
         :return:
         """
+        if not self._check_start_up_tasks():
+            self.logger.info("Startup tasks not complete - skipping loop")
+            return
         now = datetime.datetime.now()
 
         if now.hour != 19:
