+++ mongo/baseclass.py
@@ -16,6 +16,7 @@ class BaseClass(object):
     def __init__(
             self,
             ip: str = "127.0.0.1",
+            # ip: str = "ec2-35-178-51-238.eu-west-2.compute.amazonaws.com",
             username: str = None,
             password: str = None):
         """
